+++ src/routes/Card.js
@@ -23,6 +23,7 @@ module.exports = app => {
      * @apiGroup Cards 🃏️
      *
      * @apiParam (Query String Parameters) {String} [categories] Filter by categories. Multiple categories can be specified.<br/><br/> - Separate values with `,` for filtering cards containing all those categories. (AND operator)<br/> - Separate values with `|` for filtering cards containing any of those categories. (OR operator)<br/><br/>You can't mix `,` and `|`, don't insert space.<hr/>Example: `video-game,art` for cards having `video-game` AND `art` categories or `nature|animal` for cards having `nature` OR `animal` categories.
+     * @apiParam (Query String Parameters) {String} [fields] Specifies which fields to include. Each value must be separated by a `,` without space. (e.g: `label,createdAt`)
      * @apiParam (Query String Parameters) {Number{>= 1}} [limit] Limit the number of cards returned.
      * @apiUse CardResponse
      */
@@ -38,6 +39,11 @@ module.exports = app => {
                 }
                 return { $in: categories.split("|") };
             }),
+        query("fields")
+            .optional()
+            .isString().withMessage("Must be a valid string")
+            .custom(value => (/^[A-z]+(?:,[A-z]+)*$/u).test(value) ? Promise.resolve() : Promise.reject(new Error()))
+            .withMessage("Each value must be separated by a `,` without space. (e.g: `label,createdAt`)"),
         query("limit")
             .optional()
             .isInt({ min: 1 }).withMessage("Must be a valid number greater than 0.")